+++ src/client/interfaces/ifield-spec.iface.js
@@ -10,15 +10,16 @@
  *    it must be either the INPUT/SELECT element itself, or a parent of this element
  *  - dom: a CSS selector to address the value itself when 'js' is not enough or too high, or the element is special
  *    defaults to js css selector
- *  - valFrom( <item> ): a function to get the value from the provided item, defaulting to just getting the field value as `value = item[name]`
- *  - valTo( <item>, <value> ): a function to set the value into the provided item, defaulting to just setting the field value as item[name] = value
- *  - formFrom( <$node> ): a function to read the value from the form, defaulting to the 'val()' function
- *  - formTo( <$node>, <item> ): a function to write the value into the form, defaulting to the 'val( <value> )' function
- *  - check: a check function, or false (warns if unset)
- *  - type: the mandatory/optional field type, defaulting to none
+ *  - form_valFrom( <item> ): a function to get the value from the provided item, defaulting to just getting the field value as `value = item[name]`
+ *  - form_valTo( <item>, <value> ): a function to set the value into the provided item, defaulting to just setting the field value as item[name] = value
+ *  - form_formFrom( <$node> ): a function to read the value from the form, defaulting to the 'val()' function
+ *  - form_formTo( <$node>, <item> ): a function to write the value into the form, defaulting to the 'val( <value> )' function
+ *  - form_check: a check function, or false (warns if unset)
+ *  - form_type: the mandatory/optional field type, defaulting to none
  *      ref. Forms.FieldType.C
- *  - status: whether the field should be appended with an indicator to show valid|invalid state, defaulting to Checker then configured values
+ *  - form_status: whether the field should be appended with an indicator to show valid|invalid state, defaulting to Checker then configured values
  *      ref. Forms.C.ShowStatus
+ *  - form_trace: whether to be verbose for this field, defaulting to false
  */
 
 import _ from 'lodash';
@@ -46,9 +47,12 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
      * @constructor
      * @returns {IFieldSpec} the instance
      */
-    constructor( name, args ){
-        logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.IFieldSpec()', name, args );
+    constructor( args ){
+        logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.IFieldSpec()', args );
         super( ...arguments );
+
+        if( args.form_trace ) logger.debug( 'IFieldSpec()', args );
+
         return this;
     }
 
@@ -80,18 +84,27 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
 
     /**
      * @summary This is the way the status should be displayed for this field.
-     *  It is only considered if the package is configured for this way be overridable on a per-field basis.
+     *  It is only considered if the package is configured for this status be overridable on a per-field basis.
      * @returns {String} a value from Forms.C.ShowStatus, or null
      */
     iSpecStatus(){
         logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecStatus()' );
         const defn = this._defn();
-        const status = defn.form_status || null;
+        let status = defn.form_status || null;
         if( status && !Object.values( Forms.C.ShowStatus ).includes( status )){
             logger.warn( 'IFieldSpec.iSpecStatus() unexpected form_status', this.name(), status );
             status = null;
         }
         return status;
+    }
+
+    /**
+     * @returns {Boolean} whether to be verbose for this field, defaulting to false
+     */
+    iSpecTrace(){
+        logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecTrace()' );
+        const defn = this._defn();
+        return defn.form_trace === true;
     }
 
     /**
@@ -105,7 +118,8 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
 
     /**
      * @returns {String} the js css selector, or null
-     *  This UI selector is notably used for all UI operations, i.e. setting the status and validity indicators
+     *  This UI selector is notably used for all UI operations
+     *  Notably, status and validity indicators will be set around this UISelector element
      */
     iSpecUISelector(){
         logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecUISelector()' );
@@ -115,7 +129,7 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
 
     /* Maintainer note:
      *  iSpecValueFrom() get the value from the item
-        It is only used if there is not specific formTo() function.
+        It is used when there is not specific formTo() function (which should be the general case anyway).
      *  Setting (whether reactively or not) the value into the item is the responsability of checkfFn functions
      *  see iRunValueFrom() (resp. iRunValueTo()) get (resp. set) the value from (resp. to) the form
      * 
@@ -128,11 +142,13 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
      */
     iSpecValueFrom( item ){
         logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecValueFrom()' );
+        if( this.iSpecTrace()) logger.debug( 'iSpecValueFrom()', this.name(), 'item', item );
         const defn = this._defn();
         let value = null;
         if( defn.form_itemFrom ){
             assert( typeof defn.form_itemFrom === 'function', 'expect form_itemFrom() be a function, found '+defn.form_itemFrom );
             value = defn.form_itemFrom( item );
+            if( this.iSpecTrace()) logger.debug( 'iSpecValueFrom()', this.name(), 'defn.form_itemFrom() returns', value );
         } else {
             let name = this.name();
             if( name.match( /\.\$\./ )){
@@ -150,7 +166,7 @@ export const IFieldSpec = DeclareMixin(( superclass ) => class extends superclas
      *  It defaults to UI selector.
      */
     iSpecDOMSelector(){
-        logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecUISelector()' );
+        logger.verbose({ verbosity: Forms.configure().verbosity, against: Forms.C.Verbose.FUNCTIONS }, 'IFieldSpec.iSpecDOMSelector()' );
         const defn = this._defn();
         return defn.dom || this.iSpecUISelector();
     }